rename provable folder to 'lib'
