+++ lib/encoding.ts
