+++ lib/generic.ts
