+++ mina-transaction/gen/transaction.ts
@@ -17,11 +17,8 @@ import {
   SequenceState,
   ReceiptChainHash,
 } from '../transaction-leaves.js';
-import { GenericProvableExtended } from '../../provable/generic.js';
-import {
-  ProvableFromLayout,
-  GenericLayout,
-} from '../../provable/from-layout.js';
+import { GenericProvableExtended } from '../../lib/generic.js';
+import { ProvableFromLayout, GenericLayout } from '../../lib/from-layout.js';
 import * as Json from './transaction-json.js';
 import { jsLayout } from './js-layout.js';
